Page Welcome, affichage du nombre de end user crées les 30 derniers jours

+++ src/GamificationProject/src/main/java/ch/heigvd/amt/gamification/services/EndUsersManagerLocal.java
@@ -13,5 +13,5 @@ import javax.ejb.Local;
  */
 @Local
 public interface EndUsersManagerLocal {
-    public long numberOfEndUsers(int numberOfDay);
+    public int numberOfEndUsers(int numberOfDay);
 }
